[issue21953] Move tests to the aio part
* Export the new metadata type through the ``__init__.py``
* Move the tests from the old location to the new one for the
asynchronous stack.
